Move definition of GlobalEnv type to work in progress
Having a type that separates system-provided definitions from the ones
made by the user would be the cleanest solution to provide
visualizations only with data entered by the user.  However, since
Clojure makes it quite awkward to define new collections (and we have to
repeat the exercise for ClojureScript) we'll just dissociate the known
global data from the current global environment when passing data to
visualizations.